feature: комментарии реализованы 1.2(учтены замечания ревью)

+++ main-service/src/main/java/ewm/comment/service/CommentServiceImp.java
@@ -5,6 +5,7 @@ import ewm.comment.dto.CommentMapper;
 import ewm.comment.dto.NewCommentDto;
 import ewm.comment.model.Comment;
 import ewm.comment.repository.CommentRepository;
+import ewm.exceptions.BadRequestException;
 import ewm.exceptions.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
@@ -12,6 +13,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.Clock;
 import java.util.List;
 
 @Service
@@ -19,12 +21,14 @@ import java.util.List;
 @Transactional(readOnly = true)
 public class CommentServiceImp implements CommentService {
     private final CommentRepository commentRepository;
+    private final CommentMapper commentMapper;
 
     @Override
     @Transactional
     public CommentDto postComment(Long userId, Long eventId, NewCommentDto newCommentDto) {
-        Comment comment = CommentMapper.newCommentDtoToComment(newCommentDto, eventId, userId);
-        return CommentMapper.commentToDto(commentRepository.save(comment));
+        Clock clock = Clock.systemDefaultZone();
+        Comment comment = commentMapper.newCommentDtoToComment(newCommentDto, eventId, userId, clock);
+        return commentMapper.commentToDto(commentRepository.save(comment));
     }
 
     @Override
@@ -32,29 +36,35 @@ public class CommentServiceImp implements CommentService {
     public CommentDto patchComment(Long userId, Long comId, NewCommentDto newCommentDto) {
         Comment comment = commentRepository.findById(comId)
                 .orElseThrow(() -> new NotFoundException("комментарий с ИД-" + comId + " не найден"));
+        if (!comment.getAuthorId().equals(userId)) {
+            throw new BadRequestException("пользователь " + userId + " не является автором комментария " + comId);
+        }
         comment.setText(newCommentDto.getText());
         Comment updateComment = commentRepository.save(comment);
-        return CommentMapper.commentToDto(updateComment);
+        return commentMapper.commentToDto(updateComment);
     }
 
     @Override
     @Transactional
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId, Long userId) {
         if (!commentRepository.existsById(commentId)) {
             throw new NotFoundException("Comment with ID " + commentId + " not found");
+        }
+        if (!commentRepository.existsByIdAndAuthorId(commentId, userId)) {
+            throw new NotFoundException("Автор " + userId + " не является автором комментария " + commentId);
         }
         commentRepository.deleteById(commentId);
     }
 
     @Override
     public List<CommentDto> getComments(Integer from, Integer size, Long eventId) {
+        if (size.equals(0)) {
+            throw new BadRequestException("параметр size не может быть равен 0");
+        }
         int pageNumber = from / size;
         Pageable pageable = PageRequest.of(pageNumber, size);
         List<Comment> commentList = commentRepository.findAllCommentsByEventId(eventId, pageable);
-        if (commentList.isEmpty()) {
-            throw new NotFoundException("Комментарии не найдены");
-        }
-        List<CommentDto> commentDtoList = commentList.stream().map(CommentMapper::commentToDto).toList();
+        List<CommentDto> commentDtoList = commentList.stream().map(commentMapper::commentToDto).toList();
         return commentDtoList;
     }
 }
